New feature 'keyframe sets'
Keyframe sets are an optional label that can be applied to keyframes, via a new column 'Set' in the keyframes table. If a keyframe set has been provided, then the keyframe will render in a different color in the timeline (default keyframes are still black). The color used for the triangle is chosen so that all keyframes in the same set have the same color, and are unlikely to be close to a different set color.

The SetLen column shows the number of frames until the next keyframe in the same set (or 0f if it is the last keyframe in the set)