fix: editing an animated spline point that has been linked to a bone with Animation Mode off will now apply offset as expected, instead of throwing an error.
What's happening is that once a spline point has been linked to a bone, modifying it via the work area creates waypoints for all of its properties, not just the one that was modifed. The problem is that one of those properties is a hidden bool parameter, and animated bool parameters can't be modified outside of Animation Mode. Even though this hidden bool's value never changes, this bool blocks the offset from going through and you get a confusing 'You must be in Animate-Editing-Mode to directly manipulate this value' message.

The not-great fix was to treat animated bool values as if they're valid for offsets, even though they aren't, and just don't do anything with them. There might be other cases where this change might lead to a silent failure rather than an expected error message.